+++ java/client/src/main/java/glide/api/commands/ServerManagementClusterCommands.java
@@ -170,6 +170,7 @@ public interface ServerManagementClusterCommands {
 
     /**
      * Get the values of configuration parameters.<br>
+     * Starting from server version 7, command supports multiple parameters.<br>
      * The command will be sent to a random node.
      *
      * @see <a href="https://valkey.io/commands/config-get/">valkey.io</a> for details.
@@ -186,7 +187,8 @@ public interface ServerManagementClusterCommands {
     CompletableFuture<Map<String, String>> configGet(String[] parameters);
 
     /**
-     * Get the values of configuration parameters.
+     * Get the values of configuration parameters.<br>
+     * Starting from server version 7, command supports multiple parameters.
      *
      * @see <a href="https://valkey.io/commands/config-get/">valkey.io</a> for details.
      * @param parameters An <code>array</code> of configuration parameter names to retrieve values
@@ -210,6 +212,7 @@ public interface ServerManagementClusterCommands {
 
     /**
      * Sets configuration parameters to the specified values.<br>
+     * Starting from server version 7, command supports multiple parameters.<br>
      * The command will be sent to all nodes.
      *
      * @see <a href="https://valkey.io/commands/config-set/">valkey.io</a> for details.
@@ -226,7 +229,8 @@ public interface ServerManagementClusterCommands {
     CompletableFuture<String> configSet(Map<String, String> parameters);
 
     /**
-     * Sets configuration parameters to the specified values.
+     * Sets configuration parameters to the specified values.<br>
+     * Starting from server version 7, command supports multiple parameters.
      *
      * @see <a href="https://valkey.io/commands/config-set/">valkey.io</a> for details.
      * @param parameters A <code>map</code> consisting of configuration parameters and their